requirements.txt
# --------- pip ------------- #
pip==23.3.2 # Required for older version of pytorch-lightning

# --------- pytorch --------- #
torch==2.4.1
torchvision==0.19.1
pytorch-lightning==1.7.1
torchmetrics==0.10.1

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
mlflow==1.28.0

# --------- dlup ------------- #
dlup==0.3.23

# --------- others --------- #
pyrootutils~=1.0.4
pre-commit      # hooks for applying linters on commit
rich==12.5.1
pytest~=7.1.3
jupyter==1.0.0
matplotlib==3.6.0
seaborn==0.12.0
scikit-learn==1.1.2
h5py==3.7.0
omegaconf~=2.2.3
numpy~=1.23.4
scipy~=1.9.2
pandas~=1.5.0
pillow~=9.2.0
packaging~=21.3
pyvips~=2.2.1
tqdm~=4.64.1
openpyxl